+++ lib/whirly/spinners.rb
@@ -1,6 +1,7 @@
-require "json"
-
 module Whirly
-  SPINNERS = JSON.load(File.read(File.dirname(__FILE__) + "/../../data/spinners.json"))
-  SPINNERS["whirly"] = { "proc" => ->(){ [0x1F600 + rand(55)].pack("U") }, "interval" => 200 }
+  module Spinners
+  end
 end
+
+require_relative "spinners/whirly"
+require_relative "spinners/cli"