+++ lib/whirly/spinners/cli.rb
@@ -1,0 +1,7 @@
+require "json"
+
+module Whirly
+  module Spinners
+    CLI = JSON.load(File.read(File.dirname(__FILE__) + "/../../../data/cli-spinners.json")).freeze
+  end
+end