+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Different newline behaviour; append newline by default after spinner ran. Use position: "below" for old behaviour
 - Support multiple frame modes: "linear", "random", "reverse", "swing"
 - Proper unrendering (use unicode-display\_width)
+- Introduce spinner packs (to deal with eventual name conflicts, currently: whirly + cli)
+- Add more bundled spinners
 - Update CLI spinners to v0.3.0 (two new spinners)
 
 ### 0.1.1